\name{exploBATCHcolon-package}
\alias{exploBATCHcolon-package}
\alias{exploBATCHcolon}
\docType{package}
\title{
\packageTitle{exploBATCHcolon}
}
\description{
\packageDescription{exploBATCHcolon}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{exploBATCHcolon}
\packageIndices{exploBATCHcolon}
A package with colorectal cancer gene expression dataset used in Nyamundanda et al (2017) manuscript on batch detection and correction.
}
\author{
\packageAuthor{exploBATCHcolon}

Maintainer: \packageMaintainer{exploBATCHcolon}
}
\references{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A., 2017. A novel and robust statistical method to diagnose and correct batch effects in genomic data.
}
\examples{
data(Colon)  # Data for the second example in the manuscript.
}
